+++ src/data/import_from_illumina.py
@@ -32,11 +32,11 @@ def get_output_files(reportpath: str, working_dir: Path, assembly: str):
     output_dir.mkdir(exist_ok=True)
 
     # determine map outputfile. get the basename of the prefix
-    output_map = Path(reportpath).stem + "_updated.map"
+    output_map = Path(reportpath).stem + "_updated.map".replace(" ", "_")
     output_map = output_dir / output_map
 
     # creating ped file for writing updated genotypes
-    output_ped = Path(reportpath).stem + "_updated.ped"
+    output_ped = Path(reportpath).stem + "_updated.ped".replace(" ", "_")
     output_ped = output_dir / output_ped
 
     return output_dir, output_map, output_ped
@@ -74,7 +74,7 @@ def main(
     if assembly not in WORKING_ASSEMBLIES:
         raise Exception(f"assembly {assembly} not managed by smarter")
 
-    assembly_conf = WORKING_ASSEMBLIES[assembly]
+    src_assembly = WORKING_ASSEMBLIES[assembly]
 
     # custom method to check a dataset and ensure that needed stuff exists
     dataset, [snpfilepath, reportpath] = fetch_and_check_dataset(
@@ -119,8 +119,8 @@ def main(
 
     # fetch coordinates relying assembly configuration
     report.fetch_coordinates(
-        version=assembly_conf.version,
-        imported_from=assembly_conf.imported_from
+        src_assembly=src_assembly,
+        chip_name=illumina_chip.name
     )
 
     logger.info("Writing a new map file with updated coordinates")