+++ src/data/import_metadata.py
@@ -20,7 +20,7 @@ import pandas as pd
 
 from src.features.smarterdb import global_connection, Dataset
 from src.data.common import (
-    fetch_and_check_dataset, pandas_open, get_sample_species)
+    deal_with_datasets, pandas_open, get_sample_species)
 from src.features.utils import sanitize
 
 logger = logging.getLogger(__name__)
@@ -115,7 +115,7 @@ def add_metadata_by_sample(
     SampleSpecie = get_sample_species(dst_dataset.species)
 
     for index, row in data.iterrows():
-        original_id = row.get(columns["id_column"])
+        original_id = str(row.get(columns["id_column"]))
 
         # get additional columns for original_id
         locations = get_locations(row, columns, original_id)
@@ -149,8 +149,9 @@ def add_metadata_by_sample(
     help="The raw dataset file name (zip archive) in which search datafile"
 )
 @click.option(
-    '--dst_dataset', type=str, required=True,
-    help="The raw dataset file name (zip archive) in which add metadata"
+    '--dst_dataset', type=str, required=False,
+    help=("The raw dataset file name (zip archive) in which add metadata"
+          "(def. the 'src_dataset')")
 )
 @click.option('--datafile', type=str, required=True)
 @click.option(
@@ -177,17 +178,8 @@ def main(
     if metadata_column:
         logger.warning(f"Got {metadata_column} as additional metadata")
 
-    # custom method to check a dataset and ensure that needed stuff exists
-    src_dataset, [datapath] = fetch_and_check_dataset(
-        archive=src_dataset,
-        contents=[datafile]
-    )
-
-    # this will be the dataset used to define samples
-    dst_dataset, _ = fetch_and_check_dataset(
-        archive=dst_dataset,
-        contents=[]
-    )
+    src_dataset, dst_dataset, datapath = deal_with_datasets(
+        src_dataset, dst_dataset, datafile)
 
     if sheet_name and sheet_name.isnumeric():
         sheet_name = int(sheet_name)