+++ src/data/import_phenotypes.py
@@ -15,7 +15,7 @@ import pandas as pd
 
 from src.features.smarterdb import global_connection, Dataset, Phenotype, Breed
 from src.data.common import (
-    fetch_and_check_dataset, pandas_open, get_sample_species)
+    deal_with_datasets, pandas_open, get_sample_species)
 from src.features.utils import sanitize
 
 logger = logging.getLogger(__name__)
@@ -193,8 +193,9 @@ def add_phenotype_by_alias(
     help="The raw dataset file name (zip archive) in which search datafile"
 )
 @click.option(
-    '--dst_dataset', type=str, required=True,
-    help="The raw dataset file name (zip archive) in which add metadata"
+    '--dst_dataset', type=str, required=False,
+    help=("The raw dataset file name (zip archive) in which add metadata"
+          "(def. the 'src_dataset')")
 )
 @click.option('--datafile', type=str, required=True)
 @click.option('--sheet_name',
@@ -222,17 +223,8 @@ def main(src_dataset, dst_dataset, datafile, sheet_name, breed_column,
     if additional_column:
         logger.debug(f"Got {additional_column} as additional phenotype")
 
-    # custom method to check a dataset and ensure that needed stuff exists
-    src_dataset, [datapath] = fetch_and_check_dataset(
-        archive=src_dataset,
-        contents=[datafile]
-    )
-
-    # this will be the dataset used to define samples
-    dst_dataset, _ = fetch_and_check_dataset(
-        archive=dst_dataset,
-        contents=[]
-    )
+    src_dataset, dst_dataset, datapath = deal_with_datasets(
+        src_dataset, dst_dataset, datafile)
 
     if sheet_name and sheet_name.isnumeric():
         sheet_name = int(sheet_name)