+++ tests/data/test_import_from_plink.py
@@ -113,6 +113,61 @@ class TestImportFromPlink(
            new_callable=PropertyMock)
     @patch('src.features.smarterdb.Dataset.working_dir',
            new_callable=PropertyMock)
+    def test_import_from_text_plink_src_assembly(
+            self, my_working_dir, my_result_dir):
+        """Test import from plink using a source assembly"""
+
+        # create a temporary directory using the context manager
+        with tempfile.TemporaryDirectory() as tmpdirname:
+            working_dir = pathlib.Path(tmpdirname)
+            results_dir = working_dir / "results"
+
+            # assign return value to mocked property
+            my_working_dir.return_value = working_dir
+            my_result_dir.return_value = results_dir
+
+            # copy test data files
+            self.link_files(working_dir)
+
+            result = self.runner.invoke(
+                import_from_plink,
+                [
+                    "--dataset",
+                    "test.zip",
+                    "--file",
+                    "plinktest",
+                    "--chip_name",
+                    self.chip_name,
+                    "--assembly",
+                    "OAR4",
+                    "--create_samples",
+                    "--src_version",
+                    "Oar_v3.1",
+                    "--src_imported_from",
+                    "manifest"
+                ]
+            )
+
+            self.assertEqual(0, result.exit_code, msg=result.exception)
+            self.assertEqual(SampleSheep.objects.count(), 2)
+
+            # check imported chip_name attribute
+            for sample in SampleSheep.objects:
+                self.assertEqual(sample.chip_name, self.chip_name)
+
+            plink_path = results_dir / "OAR4" / "plinktest_updated"
+            plink_file = plinkfile.open(str(plink_path))
+
+            sample_list = plink_file.get_samples()
+            locus_list = plink_file.get_loci()
+
+            self.assertEqual(len(sample_list), 2)
+            self.assertEqual(len(locus_list), 3)
+
+    @patch('src.features.smarterdb.Dataset.result_dir',
+           new_callable=PropertyMock)
+    @patch('src.features.smarterdb.Dataset.working_dir',
+           new_callable=PropertyMock)
     def test_import_from_binary_plink(self, my_working_dir, my_result_dir):
         # create a temporary directory using the context manager
         with tempfile.TemporaryDirectory() as tmpdirname: