+++ tests/features/test_smarterdb.py
@@ -46,7 +46,7 @@ class CountryTestCase(MongoMockMixin, unittest.TestCase):
 
 class BreedTestCase(MongoMockMixin, unittest.TestCase):
     def setUp(self):
-        # createing a sample breed
+        # creating a sample breed
         self.breed = Breed(
             species="Sheep",
             name="Texel",
@@ -367,6 +367,17 @@ class LocationTestCase(VariantMixin, MongoMockMixin, unittest.TestCase):
             ["A", "T"]
         )
 
+    def test_no_information(self):
+        """Get an exception while querying for affymetrix in a illumina
+        record"""
+
+        self.assertRaisesRegex(
+            SmarterDBException,
+            "There's no information for",
+            self.location.is_affymetrix,
+            "T/C"
+        )
+
     def test_is_affy(self):
         for genotype in ["T/C", "T/T", "C/T", "C/C", "0/0"]:
             genotype = genotype.split("/")
@@ -376,10 +387,10 @@ class LocationTestCase(VariantMixin, MongoMockMixin, unittest.TestCase):
                 msg=f"{genotype} is not in affymetrix coordinates!"
             )
 
-        # is not in not if contains an allele not in top format
+        # this is false, for example with top coding
         for genotype in ["A/A", "A/G", "G/A", "G/G"]:
             self.assertFalse(
-                self.location.is_forward(genotype),
+                self.affy_location.is_affymetrix(genotype),
                 msg=f"{genotype} is in affymetrix coordinates!"
             )
 
@@ -413,7 +424,8 @@ class VariantSheepTestCase(VariantMixin, MongoMockMixin, unittest.TestCase):
     def test__str(self):
         self.assertEqual(
             str(self.variant),
-            "name='250506CS3900065000002_1238.1', rs_id='rs55630613'"
+            "name='250506CS3900065000002_1238.1', rs_id='['rs55630613']', "
+            "illumina_top='A/G'"
         )
 
     def test_location(self):
@@ -446,7 +458,7 @@ class VariantSheepTestCase(VariantMixin, MongoMockMixin, unittest.TestCase):
             SmarterDBException,
             "Couldn't determine a unique location for",
             self.variant.get_location,
-            version="Oar_v4.0",
+            version="Oar_v4.1",
             imported_from='SNPchiMp v.3'
         )
 
@@ -457,9 +469,42 @@ class VariantSheepTestCase(VariantMixin, MongoMockMixin, unittest.TestCase):
             SmarterDBException,
             "is not in locations",
             self.variant.get_location_index,
-            version="Oar_v4.0",
+            version="Oar_v4.1",
             imported_from='SNPchiMp v.3'
         )
+
+
+class AffyVariantTestCase(VariantMixin, MongoMockMixin, unittest.TestCase):
+    def setUp(self):
+        self.variant = VariantSheep.from_json(json.dumps(self.affy_data))
+
+    def test__str(self):
+        self.assertEqual(
+            str(self.variant),
+            "name='250506CS3900176800001_906.1', rs_id='['rs55630654']', "
+            "illumina_top='A/G'"
+        )
+
+        # with no name, str representation has affy_id
+        self.variant.name = None
+
+        self.assertEqual(
+            str(self.variant),
+            "affy_snp_id='Affx-122835222', rs_id='['rs55630654']', "
+            "illumina_top='A/G'"
+        )
+
+        # with no name, str representation has affy_id
+        self.variant.name = None
+
+    def test_probesets(self):
+        probeset = next(
+            filter(
+                lambda probeset: probeset.chip_name == 'AffymetrixAxiomOviCan',
+                self.variant.probesets
+            )
+        )
+        self.assertIn('AX-124359447', probeset.probeset_id)
 
 
 class GetSmarterIdTestCase(SmarterIDMixin, MongoMockMixin, unittest.TestCase):
@@ -614,6 +659,40 @@ class SampleSheepTestCase(SmarterIDMixin, MongoMockMixin, unittest.TestCase):
         self.assertEqual(sample.smarter_id, self.smarter_id)
         self.assertEqual(SampleSheep.objects.count(), 1)
         self.assertFalse(created)
+
+    def test_get_or_create_sample_breed(self):
+        """Test that also breed is involved in getting or creating samples"""
+
+        # creating sample first
+        sample, created = get_or_create_sample(
+            SampleSheep,
+            self.original_id,
+            self.dataset,
+            self.type_,
+            self.breed,
+            self.country,
+            self.chip_name,
+            self.sex,
+            self.alias
+        )
+
+        # calling the same function again, but with a new breed
+        sample, created = get_or_create_sample(
+            SampleSheep,
+            self.original_id,
+            self.dataset,
+            self.type_,
+            Breed.objects.get(species="Sheep", code="MER"),
+            self.country,
+            self.chip_name,
+            self.sex,
+            self.alias
+        )
+
+        self.assertIsInstance(sample, SampleSheep)
+        self.assertEqual(sample.smarter_id, "ITOA-MER-000000002")
+        self.assertEqual(SampleSheep.objects.count(), 2)
+        self.assertTrue(created)
 
     def test_get_sample_type(self):
         self.create_sample()